refactor: クラスとして切り出した
